Improvements to artist biographies.
Stripping HTML tags, and repositioning output directory so that it is
a sibling of Links / Playlists. The bios seem to be in a good state
now. There are still some tags like &uuml; which I haven't replaced
with the real thing. There must be some library function for doing
this properly.