Bug fix: tags and bio contents were not saved in the .dbm file
This should ensure that tags and biographies are saved along with
similar artist data in the .dbm file. In addition to being a attribute
of an artist, these objects are held in a dict keyed by artist ID
which is an attribute of root. The Artist objects are constructed de
novo after the library is scanned (e.g. after a refresh, as well as
the obvious case of a fresh scan), and then their {tags,bios,sim
artist data} are restored from these dicts.