New biography.update() procedure.
If the biography doesn't exist on disk, we download it and write it
together with the metadata to disk. Otherwise we read it (necessary in
order to obtain the biography, since we don't store them in RAM), and
write to disk if the new metadata differs from that on disk.