Improve information for user at library_scan and create_links_... entry points.
If the user selects "Scan Library" she is given the option of
instructing the program to go straight on to the downloads when the
scan is finished.

If last.fm downloads (and content creation) are going to be done, then
the user is forced to make a final review of the settings.